+++ runners/runner_asset_allocation.py
@@ -51,6 +51,7 @@ class Runner(loader.Runner):
         self.iter_msci()
         self.iter_prices()
         self.iter_fundamentals()
+        self.iter_factors()
         self.iter_marketmuliples()
         self.iter_holdings()
         self.iter_securities()
@@ -99,6 +100,7 @@ class Runner(loader.Runner):
             current_multiples = self.marketmultiple_datasource.outgoing_row(date)
             current_fundamentals = self.fundamentals_datasource.outgoing_row(date)
             index_components = self.portfolio_datasource.outgoing_row(date)
+            factors = self.factor_datasource.outgoing_row(date)
 
             assign_dict = {
                 "date": date,
@@ -119,7 +121,7 @@ class Runner(loader.Runner):
             # assign returns to strategies and backtest
             for strat, strat_obj in self.strategies.items():
                 strat_obj.assign(**assign_dict)
-                strat_obj.backtest(date, current_fundamentals["marketcap"])
+                strat_obj.backtest(date, current_fundamentals["marketcap"], factors)
 
         # assign weights to security objects
         for strat, strat_obj in self.strategies.items():